+++ src/main/java/vn/aptech/beehub/security/services/RefreshTokenService.java
@@ -28,7 +28,7 @@ public class RefreshTokenService {
         return refreshTokenRepository.findByToken(token);
     }
 
-    public RefreshToken createRefreshToken(Long userId) {
+    public RefreshToken createRefreshToken(int userId) {
         RefreshToken refreshToken = new RefreshToken();
 
         refreshToken.setUser(userRepository.findById(userId).get());
@@ -49,7 +49,7 @@ public class RefreshTokenService {
     }
 
     @Transactional
-    public int deleteByUserId(Long userId) {
+    public int deleteByUserId(int userId) {
         return refreshTokenRepository.deleteByUser(userRepository.findById(userId).get());
     }
 }
